Fix logfilename reference to prevent error
Looks like a copy/paste error causes this to reference a node "memcached-chat" that doesn't actually exist.  It causes the error:

"undefined method `[]' for nil:NilClass"

Not worth my effort to accept your rights-assignments document so feel free to reject this pull request and remove the 5 characters yourself (I'd have just filed a bug but you have that feature disabled).