Allow full path of log file to be spec-ed.
The default of /var/log is nice, but it is better to be able to place
the log where the admin desires.  This PR handles only the case where
the template memcached.conf is invoked.  It tries to create the directory
if it does not exist with sensible defaults of the user/group of
the memcache process.