Update default.rb
Added ":start" action for memcached service, to fix the scenario in openstack cookbbook:
In openstack, identity service depends on memcached service. But after the memcached package was installed, before service was started, the identity service started serving already, this caused the identity service can not work properly. In this case, we should make ensure the memcached service starts immediately.