Fix Memcached user on Debian and/or Ubuntu, or add if missing.
This is to fix the default user on Debian and/or Ubuntu which used to be the
`nobody` user, but that since has changed, as per:

  - https://bugs.debian.org/cgi-bin/bugreport.cgi?bug=587797
  - https://launchpad.net/ubuntu/+source/memcached/1.4.13-0.1ubuntu1

Nowadays, packages on both Debian and Ubuntu (and most likely their derivatives)
would create dedicated `memcache` user. Nonetheless, we attempt to create
a dedicated user in case there isn't one already.

Signed-off-by: Krzysztof Wilczynski <krzysztof.wilczynski@linux.com>