Initial conversion to a true library cookbook
Convert the definition to a 12.5+ custom resource and use
compat_resource for 12.X compatibility.  This removes the need to use
the definition to setup instance values while also setting config in
node attributes.  You can use the memcached_instance custom resource
without applying a thing on the node (recipe or attributes).