Prevent init confusion
If we have both an upstart script and a sys-v script named memcached
then service status seems to get confused as it tries to use the sys-v
script. So if we name our instance memcached we need to remove the sys-v
script. If not we can leave it there just in case someone wants to use
it.

Signed-off-by: Tim Smith <tsmith@chef.io>