Add 'no-restart' property
- 'no_restart' property aims to prevent restart of memcached
on configuration change (default to false for compatibility).
The restart is often unwanted as it empties the cache; it can
be unnecessary (example: if we fix a comment...) Often we want
to delay or orchestrate it, in particular with a cluster of
memcached (we do not want to restart all nodes at the same time).

Signed-off-by: Cyril Martin <c.martin@criteo.com>