Add 'binary-path' propertie
- 'binary-path' property aims to give more control on the
installation. When 'binary_path' is defined, it is used to
configure memcached services and the package installation
is skipped.
- commit b2e98e6 changed how the user was created. It said:
"If the user gets removed later on we should be recreating it".
But it causes an issue: we cannot upgrade this cookbook while
keeping memcached process up, because we cannot update the HOME
directory of a user executing a program (and obviously we do not
want to restart memcached service for that...)
The 'binary_path' feature can be a workaround.

Signed-off-by: Cyril Martin <c.martin@criteo.com>